Reorg typing funcs
